upgrade openbrush 4 with latest rust stable toolchain
