+++ ink/contracts/lucky_raffle/Cargo.toml
@@ -1,0 +1,38 @@
+[package]
+name = "lucky_raffle"
+version = "0.2.0"
+authors = ["guigou"]
+edition = "2021"
+
+[dependencies]
+ink = { version = "4.3.0", default-features = false}
+scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
+scale-info = { version = "2.9.0", default-features = false, features = ["derive"], optional = true }
+openbrush = { version = "4.0.0-beta", git = "https://github.com/Brushfam/openbrush-contracts", default-features = false, features = ["access_control"] }
+lucky = { path = "../../logics", default-features = false }
+random_generator = { path = "../random_generator", default-features = false, features = ["ink-as-dependency"] }
+reward_manager = { path = "../reward_manager", default-features = false, features = ["ink-as-dependency"] }
+dapps_staking_developer = { path = "../dapps_staking_developer", default-features = false, features = ["ink-as-dependency"] }
+
+[lib]
+path = "lib.rs"
+
+[features]
+default = ["std"]
+std = [
+    "ink/std",
+    "scale/std",
+    "scale-info/std",
+    "openbrush/std",
+    "lucky/std",
+    "random_generator/std",
+    "reward_manager/std",
+    "dapps_staking_developer/std",
+]
+ink-as-dependency = []
+
+[profile.release]
+overflow-checks = false
+
+[profile.dev]
+overflow-checks = false
