+++ ink/logics/lib.rs
@@ -1,0 +1,2 @@
+#![cfg_attr(not(feature = "std"), no_std, no_main)]
+pub mod traits;
