+++ ink/logics/traits/participant_filter/mod.rs
@@ -1,0 +1,2 @@
+pub mod filter_latest_winners;
+pub mod participant_filter;
