--- logics/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "lucky"
-version = "0.1.0"
-authors = ["guigou"]
-edition = "2021"
-
-[dependencies]
-ink = { version = "4.2.0", default-features = false}
-scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
-scale-info = { version = "2.6.0", default-features = false, features = ["derive"], optional = true }
-openbrush = { version = "3.0.0", git = "https://github.com/727-Ventures/openbrush-contracts", default-features = false, features = ["access_control"]}
-
-[lib]
-name = "lucky"
-path = "lib.rs"
-crate-type = ["rlib"]
-
-[features]
-default = ["std"]
-std = [
-    "ink/std",
-    "scale/std",
-    "scale-info/std",
-    "openbrush/std",
-]
-ink-as-dependency = []
