--- logics/impls/participant_filter/mod.rs
@@ -1,1 +1,0 @@
-pub mod filter_latest_winners;
