--- logics/lib.rs
@@ -1,5 +1,0 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-#![feature(min_specialization)]
-
-pub mod traits;
-pub mod impls;
