--- logics/mod.rs
@@ -1,3 +1,0 @@
-pub mod traits;
-pub mod impls;
-
