--- logics/traits/mod.rs
@@ -1,6 +1,0 @@
-pub mod participant_manager;
-pub mod participant_filter;
-pub mod random;
-pub mod random_generator;
-pub mod reward;
-pub mod raffle;
