--- logics/traits/participant_filter/mod.rs
@@ -1,2 +1,0 @@
-pub mod participant_filter;
-pub mod filter_latest_winners;
