--- logics/traits/reward/mod.rs
@@ -1,1 +1,0 @@
-pub mod psp22_reward;
